show current action (helpful for big uploads)
The problem is that "published ..." message is there only when upload is done.
Uploading big artifact to repository can take significant amount of time. It can take minutes.
Or even dozens of minutes for big uploads over slow Internet uplink.

It is not obvious what happens when upload is started. The console just hangs.
More user friendly would be to show for a user what is about to happen. In this case upload of an artifact will take place.